Post processing effects transferred from the player to the trigger volume in the scene
Copied post processing profile to trigger volume in the main menu scene. The player has a blank profile that acts as the default post processing settings and the trigger volume overwrites those settings while the player is within the volume.

Added physics to the props on the deck in the main menu scene.