Score system added
Scoring system added, player can collect items by grabbing them and pressing the activate button "trigger" or if the object is set to be collected on a collision event, such as shooting, throwing something at the object, or if the player collides with it. Score is displayed in the wrist menu.